refactor: rename guard
