+++ app/frontend/src/app/modules/app/app.module.ts
@@ -10,13 +10,13 @@ import { StyleClassModule } from 'primeng/styleclass';
 import AppRoutingModule from 'src/app/modules/app/app-routing.module';
 import AppComponent from 'src/app/modules/app/components/app/app.component';
 import AuthModule from 'src/app/modules/auth/auth.module';
-import { AuthGuardService } from 'src/app/modules/auth/services/auth-guard.service';
+import { AuthInterceptor } from 'src/app/modules/auth/interceptors/auth.interceptor';
+import { AuthGuardService } from 'src/app/modules/auth/services/auth.guard';
 import CoursesModule from 'src/app/modules/courses/courses.module';
 import FooterComponent from 'src/app/shared/components/footer/footer.component';
 import HeaderComponent from 'src/app/shared/components/header/header.component';
 import LoadingComponent from 'src/app/shared/components/loading/loading.component';
 import NotFoundComponent from 'src/app/shared/components/not-found/not-found.component';
-import { AuthInterceptor } from 'src/app/modules/auth/interceptors/auth.interceptor';
 
 @NgModule({
   imports: [
